rename chat chat-client
