+++ lib/index.js
@@ -1,6 +1,6 @@
 // @flow
 import Service, {type InitOptions} from './service'
-import ChatClient from './chat'
+import ChatClient from './chat-client'
 import BotInfo from './utils/keybaseStatus'
 
 class Bot {
