+++ lib/service/index.js
@@ -7,7 +7,7 @@ export type InitOptions = {|
   verbose?: boolean,
 |}
 
-class BaseBot {
+class Service {
   initializedYet: boolean
   username: void | string
   devicename: void | string
@@ -21,7 +21,7 @@ class BaseBot {
 
   /**
    * Initializes your bot.
-   * @memberof BaseBot
+   * @memberof Service
    */
   async init(username: string, paperkey: string, options?: InitOptions): Promise<void> {
     if (!username || typeof username !== 'string') {
@@ -53,7 +53,7 @@ class BaseBot {
 
   /**
    * Deinitializes the bot by removing its own directory and shutting down.
-   * @memberof BaseBot
+   * @memberof Service
    */
   async deinit(): Promise<void> {
     await keybaseExec(this.homeDir, ['logout'])
@@ -63,7 +63,7 @@ class BaseBot {
 
   /**
    * Returns the username and device name of the current bot, or `null` if the bot has not been initialized.
-   * @memberof BaseBot
+   * @memberof Service
    */
   myInfo(): ?BotInfo {
     if (this.username && this.devicename && this.homeDir) {
@@ -96,4 +96,4 @@ class BaseBot {
   }
 }
 
-export default BaseBot
+export default Service
